refactor(tui): complete workspace directory and thin the facade (W5)
status.ts and view.tsx move into the workspace domain; app.tsx and the
component/unit suites now import the public facade from ./workspace. The root
workspace-controller.ts compat re-export is deleted. Legacy workspace-* test
files move to tests/unit/tui/workspace/ with domain names so the stale-path
pass is genuinely clean. No behavior change; 80x24 and 120x32 PTY smokes pass.
