+++ src/tui/app.tsx
@@ -65,7 +65,7 @@ import { createInputQueue } from "./input-queue";
 import { routeCommandSubmission } from "./command-scheduler";
 import { createSideQuestionController } from "./side-question-controller";
 import { createSkillPickerController } from "./skill-picker-controller";
-import { createWorkspaceController } from "./workspace-controller";
+import { createWorkspaceController } from "./workspace";
 import { createQueuedWorkController } from "./queued-work-controller";
 import { createStageSessionController } from "./stage-session-controller";
 import { createStartupController } from "./startup-controller";
@@ -73,7 +73,7 @@ import { activateSkillForSession } from "../core/skills";
 import { initializeSessionIdentity, type SessionIdentity } from "../core/agent-loop/session-init";
 import { createSessionIdentityCoordinator } from "./session-identity-coordinator";
 import { SideQuestionOverlay } from "./views/SideQuestionOverlay";
-import { WorkspacePage } from "./views/WorkspacePage";
+import { WorkspacePage } from "./workspace/view";
 import { copyTextToClipboard } from "./clipboard";
 import { closeAllProviderSessions, closeProviderSession } from "../providers/lifecycle";
 import { registerHostShutdownCleanup } from "../core/process/shutdown";
