+++ tests/unit/tui/workspace/buffer-io.test.ts
@@ -8,9 +8,9 @@ import {
   isEditablePreview,
   readEditableFile,
   readFileStamp,
-} from "../../../src/tui/workspace/buffer-io";
-import { assertProjectRelativePath } from "../../../src/tui/workspace/paths";
-import { readFilePreview } from "../../../src/tui/workspace/tree-data";
+} from "../../../../src/tui/workspace/buffer-io";
+import { assertProjectRelativePath } from "../../../../src/tui/workspace/paths";
+import { readFilePreview } from "../../../../src/tui/workspace/tree-data";
 
 let root: string;
