+++ tests/unit/tui/workspace/editor-controller.test.ts
@@ -2,9 +2,9 @@ import { afterEach, beforeEach, describe, expect, test } from "bun:test";
 import { mkdtemp, mkdir, rename, rm, stat, utimes, writeFile } from "node:fs/promises";
 import { join } from "node:path";
 import { tmpdir } from "node:os";
-import { createWorkspaceController } from "../../../src/tui/workspace-controller";
-import type { TuiKeyEvent } from "../../../src/tui/decision-interaction";
-import { eventually } from "../../support/async/eventually";
+import { createWorkspaceController } from "../../../../src/tui/workspace";
+import type { TuiKeyEvent } from "../../../../src/tui/decision-interaction";
+import { eventually } from "../../../support/async/eventually";
 import type { TextareaRenderable } from "@opentui/core";
 
 let root: string;
