+++ tests/unit/tui/workspace/file-operations.test.ts
@@ -10,7 +10,7 @@ import {
   moveEntry,
   removeFile,
   trashEntry,
-} from "../../../src/tui/workspace/file-operations";
+} from "../../../../src/tui/workspace/file-operations";
 
 let root: string;
